services/tracing/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-jaeger-thrift==1.20.0
jaeger-client==4.8.0
redis==5.0.1
psycopg2-binary==2.9.9
